```
$ go test -v -run none -bench . -benchmem -benchtime 10s                                                                                      1m 40s 916ms
goos: linux
goarch: amd64
pkg: github.com/ipfs/go-ds-crdt
BenchmarkSync/1-8 	      10	1808648650 ns/op	992140285 B/op	 1747320 allocs/op
BenchmarkSync/2-8 	      10	1788246614 ns/op	992006935 B/op	 1745158 allocs/op
BenchmarkSync/3-8 	      10	1792477519 ns/op	992163355 B/op	 1748719 allocs/op
BenchmarkSync/4-8 	      10	1814152113 ns/op	992217028 B/op	 1749633 allocs/op
BenchmarkSync/5-8 	      10	1787339944 ns/op	992124135 B/op	 1747595 allocs/op
BenchmarkSync/10-8         	      10	1779437564 ns/op	992187234 B/op	 1749065 allocs/op
BenchmarkSync/15-8         	      10	1796560275 ns/op	992003196 B/op	 1745018 allocs/op
BenchmarkSync/30-8         	      10	1809471334 ns/op	992041555 B/op	 1745974 allocs/op
BenchmarkSync/50-8         	      10	1808551099 ns/op	992190058 B/op	 1749196 allocs/op
PASS
ok  	github.com/ipfs/go-ds-crdt	187.641s
```